working on delete route still

+++ app/api/watchlist_stocks_routes.py
@@ -19,26 +19,26 @@ watchlist_stocks_routes = Blueprint('watchlist_stocks', __name__)
 def watchlist():
     watchlist_stocks = WatchlistStocks.query.filter(
         WatchlistStocks.user_id == current_user.id).all()
-    print('>>>>>>>>>>>', watchlist_stocks)
-    return {[stock.to_dict() for stock in watchlist_stocks]} # removed key of 'watchlist'-215pm07/20/21
+    return {'watchlist':[stock.to_dict() for stock in watchlist_stocks]}
 
 
 
 # make this a post
-@watchlist_stocks_routes.route('/<ticker>', methods= ['GET','POST'])
-@login_required
-def add_to_watchlist():
-    new_addition = WatchlistStocks.query.filter(
-        WatchlistStocks.ticker == current_user.ticker).one()
-    return {[ticker.to_dict() for ticker in new_addition]}
+# @watchlist_stocks_routes.route('/<ticker>', methods= ['GET','POST'])
+# @login_required
+# def add_to_watchlist():
+#     new_addition =
 
 
 
-# ask about id vs ticker(is get necessary here)
+
 @watchlist_stocks_routes.route('/<ticker>', methods=['GET', 'DELETE'])
+@login_required
 def delete_watchlist_ticker(ticker):
-    delete_stock = WatchlistStocks.query.get(ticker)
-    db.session.delete(delete_stock)
-    db.session.commit()
+    if request.method == 'DELETE':
+        delete_stock = WatchlistStocks.query.filter(
+            WatchlistStocks.user_id == current_user.id and WatchlistStocks.ticker == ticker).first()
+        db.session.delete(delete_stock)
+        db.session.commit()
 
-    return delete_stock.to_dict()
+    return delete_stock.to_dict() # not sure what to do here yet(want to return watchlist minus ticker deletion)
